requirements.txt
django
django-rest-framework
pillow
django-cors-headers